将其中的加权数类型由int 改为了double

+++ 银行家算法模拟/银行家算法模拟/Safe_Sequence_Sort.h
@@ -17,7 +17,7 @@ public:
 	void WriteOfSafeAndWeight();
 	void Run();
 private:
-	vector<pair<string, int> > m_SafeAndWeight;
+	vector<pair<string, double> > m_SafeAndWeight;
 	ifstream m_ReadSafeSequencetxt;
 	ofstream m_WriteSafeSequence;
 };
